+++ spring-integration-file/src/main/java/org/springframework/integration/file/remote/aop/RotationPolicy.java
@@ -17,17 +17,18 @@
 package org.springframework.integration.file.remote.aop;
 
 import org.springframework.integration.core.MessageSource;
+import org.springframework.util.Assert;
 
 /**
- * Implementations can reconfigure the message source before and/or after
- * a poll.
+ * A strategy for rotating advices to allow reconfiguring
+ * the message source before and/or after a poll.
  *
  * @author Gary Russell
  * @author Michael Forstner
  * @author Artem Bilan
  * @author David Turanski
  *
- * @since 5.0.7
+ * @since 5.2
  */
 public interface RotationPolicy {
 
@@ -44,11 +45,42 @@ public interface RotationPolicy {
 	 */
 	void afterReceive(boolean messageReceived, MessageSource<?> source);
 
-
 	/**
-	 *
+	 * Return the current {@link KeyDirectory}.
 	 * @return the current {@link KeyDirectory}
+	 * @since 5.2
 	 */
 	KeyDirectory getCurrent();
+
+	/**
+	 * A key for a thread-local store and its related directory pair.
+	 */
+	class KeyDirectory {
+
+		private final Object key;
+
+		private final String directory;
+
+		public KeyDirectory(Object key, String directory) {
+			Assert.notNull(key, "key cannot be null");
+			Assert.notNull(directory, "directory cannot be null");
+			this.key = key;
+			this.directory = directory;
+		}
+
+		public Object getKey() {
+			return this.key;
+		}
+
+		public String getDirectory() {
+			return this.directory;
+		}
+
+		@Override
+		public String toString() {
+			return "KeyDirectory [key=" + this.key.toString() + ", directory=" + this.directory + "]";
+		}
+
+	}
 
 }
