+++ src/reference/asciidoc/whats-new.adoc
@@ -55,6 +55,13 @@ See <<./aggregator.adoc#flux-aggregator,Flux Aggregator>> for more information.
 The FTP and SFTP modules now provide an event listener for certain Apache Mina FTP/SFTP server events.
 See <<./ftp.adoc#ftp-server-events, Apache Mina FTP Server Events>> and <<./sftp.adoc#sftp-server-events, Apache Mina SFTP Server Events>> for more information.
 
+[[x5.2-avro]]
+==== Avro Transformers
+
+Simple Apache Avro transformers are now provided.
+See <<./transformers.adoc#avro-transformers, Avro Transformers>> for more information.
+
+
 [[x5.2-general]]
 === General Changes
 
@@ -131,14 +138,14 @@ See <<./webflux.adoc#webflux,WebFlux Support>> for more information.
 The `MongoDbMessageStore` can now be configured with custom converters.
 See <<./mongodb.adoc#mongodb, MongoDB Support>> for more information.
 
-[[x5.2-avro]]
-==== Avro Transformers
-
-Simple Apache Avro transformers are now provided.
-See <<./transformers.adoc#avro-transformers, Avro Transformers>> for more information.
-
 [[x5.2-routers]]
 ==== Router Changes
 
 You can now disable falling back to the channel key as the channel bean name.
 See <<./router.adoc#dynamic-routers, Dynamic Routers>> for more information.
+
+[[x5.2--ftp-sftp]]
+==== FTP/SFTP Changes
+
+The `RotatingServerAdvice` is decoupled now from the `RotationPolicy` and its `StandardRotationPolicy`.
+See <<./ftp.adoc#ftp-rotating-server-advice, Polling Multiple Servers and Directories>> for more information.
